+++ packages/gov-ui-kit/CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   instead
 - Update `IProposalActionUpdateMetadataDaoMetadata` interface `logo` property to `avatar` to better align with actions.
 - Bump `softprops/action-gh-release` from 2.2.0 to 2.2.1
+- Update minor and patch NPM dependencies
 
 ## [1.0.62] - 2025-01-08