+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.0.63] - 2025-01-15
+
 ### Changed
 
 - Update `InputFileAvatar` core component to remove `onFileSelect` and `onFileError` properties and use `value` and